backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
requests==2.31.0

# Firebase for push notifications
firebase-admin==6.2.0

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
